remove legacy class
