+++ great_expectations/data_context/store/__init__.py
@@ -5,7 +5,6 @@ from .store_backend import (  # isort:skip
     StoreBackend,
 )
 from .gx_cloud_store_backend import GXCloudStoreBackend  # isort:skip
-from .ge_cloud_store_backend import GeCloudStoreBackend  # isort:skip
 from .tuple_store_backend import (  # isort:skip
     TupleFilesystemStoreBackend,
     TupleGCSStoreBackend,
